used for fox radar new method from Circle class; still need to do refinements for circle pixels;

+++ src/main/java/org/game/towers/workers/Algorithms/MathAlgorithms/MathAlgorithms.java
@@ -9,7 +9,7 @@ public class MathAlgorithms {
     public MathAlgorithms() {}
 
     public static HashMap<Coordinates, Integer> getInscribedCoordinates(double x, double y, int radius) {
-        Coordinates coordinatesCenter = new Coordinates((int)x, (int)y);
+        Coordinates coordinatesCenter = new Coordinates(x, y);
         Circle circle = new Circle(coordinatesCenter, radius);
 
         return circle.getInscribedCoordinates();
